FIX: CertService 리팩토링 - 메서드 네임 정리 , 사용 안하는 메서드 정리 - Repository 의존성 역전 적용 -> 추후 도메인 분리까지 할 예정 - QueryDSL 적용 및 동적 쿼리 작성
TODO
: JPA 의존성을 끊음으로써 자동 업데이트 기능 제한 유의해야함.
: oneToMany같은 경우 join -> 직접 조회 및 적용으로 변경 필요

+++ src/main/java/com/delgo/reward/dto/map/OtherMapDTO.java
@@ -19,10 +19,10 @@ public class OtherMapDTO {
     private int viewCount;
 
     List<CertByMungpleResDTO> certs;
-    int totalCount;
+    long totalCount;
 
 
-    public OtherMapDTO(User user, List<CertByMungpleResDTO> certs, int totalCount){
+    public OtherMapDTO(User user, List<CertByMungpleResDTO> certs, long totalCount){
         userId = user.getUserId();
         nickname = user.getName();
         viewCount = user.getViewCount();
